Separated `gen/label.go` into multiple files
