+++ gen/label_manager.go
@@ -1,0 +1,8 @@
+package gen
+
+import "alon.kr/x/usm/core"
+
+type LabelManager interface {
+	GetLabel(name string) *LabelInfo
+	NewLabel(info LabelInfo) core.Result
+}
